+++ components/mui-generic-date/index.js
@@ -1,12 +1,15 @@
-import React from 'react';
+import React, { useId } from 'react';
 import _ from 'lodash';
 import FormControl from '@mui/material/FormControl';
 import FormHelperText from '@mui/material/FormHelperText';
-import dayjs from 'dayjs';
+import FormLabel from '@mui/material/FormLabel';
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { DateTimePicker } from '@mui/x-date-pickers/DateTimePicker';
+import dayjs from 'dayjs';
 
-import { isValidDayjsFormat, passRest } from '../../helpers';
+import { isValidDayjsFormat } from '../../helpers/is-valid-dayjs-format';
+import { passRest } from '../../helpers/pass-rest';
+import { makeWidthStyle } from '../../helpers';
 
 // DOC: https://mui.com/x/api/date-pickers/date-picker/
 //      https://mui.com/x/react-date-pickers/adapters-locale/
@@ -20,29 +23,26 @@ const MuiGenericDate = ({
   onBlur,
   error,
   className,
-  disabled = false,
+  size,
+  variant,
+  color,
   readOnly = false,
   required,
   fullWidth = false,
-  variant,
   placeholder,
+  floatingLabel,
   width,
-  disableFuture,
-  disableHighlightToday,
-  disableOpenPicker,
-  disablePast,
   displayWeekNumber,
   format,
   maxDate,
   minDate,
-  reduceAnimations,
-  showDaysOutsideCurrentMonth,
   lfLocale,
   datetime = false,
   views,
   component,
   ...rest
 }) => {
+  const controlId = useId();
   const localeCode = lfLocale && _.isString(lfLocale) ? lfLocale.split('-')[0] : undefined;
   const Component = datetime ? DateTimePicker : DatePicker;
 
@@ -64,29 +64,31 @@ const MuiGenericDate = ({
         required={required}
         error={error != null}
         fullWidth={fullWidth}
+        style={makeWidthStyle(fullWidth, width)}
       >
+        {label && !floatingLabel && (
+          <FormLabel htmlFor={controlId}>{label}</FormLabel>
+        )}
         <Component
+          id={controlId}
+          required={floatingLabel ? required : undefined}
+          label={floatingLabel ? label : undefined}
           placeholder={placeholder}
+          slotProps={{ textField: {
+            size: size ?? undefined,
+            variant: variant ?? undefined,
+            color: color ?? undefined,
+            required: !!(floatingLabel && required)
+          }}}
           readOnly={readOnly}
-          required={required}
-          style={_.isNumber(width) && !fullWidth ? { width: `${parseInt(width, 10)}px` } : undefined}
           onChange={onChange}
           defaultValue={defaultValue}
-          variant={variant ?? undefined}
           onBlur={onBlur}
-          disabled={disabled}
-          label={label}
           views={!_.isEmpty(views) ? views : ['day', 'year']}
-          disableFuture={disableFuture}
-          disableHighlightToday={disableHighlightToday}
-          disableOpenPicker={disableOpenPicker}
-          disablePast={disablePast}
           displayWeekNumber={displayWeekNumber}
           format={isValidDayjsFormat(format) ? format : undefined}
           maxDate={maxDate && dayjs(maxDate)}
           minDate={minDate && dayjs(minDate)}
-          reduceAnimations={reduceAnimations}
-          showDaysOutsideCurrentMonth={showDaysOutsideCurrentMonth}
           {...passRest(rest)}
         />
         {hint && !error && <FormHelperText>{hint}</FormHelperText>}